PBI 31-36

+++ Backend/src/main/java/oasip/backend/Service/EventService.java
@@ -1,19 +1,34 @@
 package oasip.backend.Service;
 
+import oasip.backend.Config.Jwts.JwtUserDetailsService;
 import oasip.backend.DTOs.Event.EventCreateDto;
 import oasip.backend.DTOs.Event.EventDetailDto;
 import oasip.backend.DTOs.Event.EventEditDto;
 import oasip.backend.DTOs.Event.EventListAllDto;
 import oasip.backend.Enitities.Event;
+import oasip.backend.Enitities.EventCategoriesOwner;
+import oasip.backend.Enitities.Eventcategory;
+import oasip.backend.Enitities.User;
+import oasip.backend.Enum.Role;
+import oasip.backend.Exception.ErrorResponse;
 import oasip.backend.ListMapper;
+import oasip.backend.Service.Email.EmailService;
+import oasip.backend.repositories.CategoryRepository;
+import oasip.backend.repositories.EventCategoriesOwnerRepository;
 import oasip.backend.repositories.EventRepository;
+import oasip.backend.repositories.UserRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.time.format.DateTimeFormatter;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -27,16 +42,46 @@ public class EventService {
     private ListMapper listMapper;
     @Autowired
     private EventCategoryService categoryService;
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    JwtUserDetailsService jwtUserDetailService;
+    @Autowired
+    private EmailService emailService;
+    @Autowired
+    private EventCategoriesOwnerRepository eventCategoriesOwnerRepository;
+    @Autowired
+    private CategoryRepository categoryRepository;
 
-    public List<EventListAllDto> getAllEvent() {
-        List<Event> events = eventRepository.findAll(Sort.by("eventStartTime").descending());
-        return listMapper.maplist(events, EventListAllDto.class, modelMapper);
+    public ResponseEntity<?> getAllEvent() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Role role = (Role) SecurityContextHolder.getContext().getAuthentication().getAuthorities().toArray()[0];
+//        System.out.println(SecurityContextHolder.getContext().getAuthentication().getName()); // anonymousUser
+        List<Event> events;
+        if(role.getAuthority().contains("admin")) {
+            events = eventRepository.findAll(Sort.by("eventStartTime").descending());
+        }else if(role.getAuthority().contains("student")){
+            events = eventRepository.findByBookingEmail(authentication.getName(),Sort.by("eventStartTime").descending());
+        }else {
+            List<Integer> eventCategoriesOwner = eventCategoriesOwnerRepository.findAllByLecturer_Email(authentication.getName());
+            events = eventRepository.findAllByEventCategory_IdList(eventCategoriesOwner,Sort.by("eventStartTime").descending());
+        }
+        return ResponseEntity.ok(listMapper.maplist(events, EventListAllDto.class, modelMapper));
     }
 
-    public EventDetailDto getEvent(Integer eventId) {
+    public ResponseEntity<?> getEvent(Integer eventId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Role role = (Role) SecurityContextHolder.getContext().getAuthentication().getAuthorities().toArray()[0];
+        List<Integer> eventCategoriesOwner = null;
+        if(role.getAuthority().contains("lecturer")){
+            eventCategoriesOwner = eventCategoriesOwnerRepository.findAllByLecturer_Email(authentication.getName());
+        }
         Event event = eventRepository.findById(eventId).orElseThrow(
                 () -> new ResponseStatusException( HttpStatus.NOT_FOUND , eventId + " Does not Exist !!!"));
-        return modelMapper.map(event, EventDetailDto.class);
+        if(role.getAuthority().contains("admin") || authentication.getName().contains(event.getBookingEmail()) || eventCategoriesOwner.contains(event.getEventCategory().getId())){
+            return ResponseEntity.ok(modelMapper.map(event, EventDetailDto.class));
+        }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponse(HttpStatus.FORBIDDEN,"Access denied"));
     }
 
     public List<EventListAllDto> filterEvents(Integer categoryId , Integer optionId , Date time) {
@@ -108,46 +153,72 @@ public class EventService {
         return listMapper.maplist(events, EventListAllDto.class, modelMapper);
     }
 
-    public EventCreateDto createEvent(EventCreateDto newEvent) {
-//        Validations validation = new Validations();
-//        if (newEvent.getEventDuration() == null) {
-//            if (newEvent.getEventCategoryId() != null) {
-//                CreateEventcategoryDto eventcategory = categoryService.getCategory(newEvent.getEventCategoryId());
-//                newEvent.setEventDuration(eventcategory.getEventCategoryDuration());
-//                validation.overlab(getEachEventCategories(newEvent.getEventCategoryId()), newEvent.getEventStartTime(), newEvent.getEventDuration());
-//            }
-//        } else {
-//            validation.overlab(getEachEventCategories(newEvent.getEventCategoryId()), newEvent.getEventStartTime(), newEvent.getEventDuration());
-//        }
-//        if (validation.getTextError().length() > 0) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, validation.getTextError());
-//        }
-        Event event = modelMapper.map(newEvent, Event.class);
-        eventRepository.saveAndFlush(event);
-        return newEvent;
+    public ResponseEntity<?> createEvent(EventCreateDto newEvent) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Role role = null;
+        if(!authentication.getName().contains("anonymousUser")){
+            role = (Role) SecurityContextHolder.getContext().getAuthentication().getAuthorities().toArray()[0];
+        }
+//        System.out.println(SecurityContextHolder.getContext().getAuthentication().getName()); // anonymousUser
+        if(authentication.getName().contains("anonymousUser") || authentication.getName().contains(newEvent.getBookingEmail()) || role.getAuthority().contains("admin")){
+            Event event = modelMapper.map(newEvent, Event.class);
+            eventRepository.saveAndFlush(event);
+            Eventcategory eventcategory = categoryRepository.getById(newEvent.getEventCategoryId());
+            Date endDate = new Date(event.getEventStartTime().getTime() + (newEvent.getEventDuration() * 60000));
+            // send Email
+            String subject = "[OASIP] " + eventcategory.getEventCategoryName() + " @ " + event.getEventStartTime() + " - " +  endDate;
+            String body = "Booking : " + event.getBookingEmail() + "\n" +
+                    "Event Category: " + eventcategory.getEventCategoryName() + "\n" +
+                    "When: " + event.getEventStartTime() + "\n" +
+                    "Event Notes: " + ((event.getEventNotes() != null) ? event.getEventNotes():" ");
+//            System.out.println(subject);
+//            System.out.println(body);
+            String status = emailService.sendSimpleMail(event.getBookingEmail() , body , subject);
+        }else{
+            if (role.getAuthority().contains("lecturer"))
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponse(HttpStatus.FORBIDDEN,"Access denied"));
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ErrorResponse(HttpStatus.BAD_REQUEST,"The booking email must be the same as the student's email"));
+        }
+        return ResponseEntity.ok(newEvent);
     }
 
     public void deleteEvent(Integer eventId) {
-        eventRepository.findById(eventId).orElseThrow(
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Role role = (Role) SecurityContextHolder.getContext().getAuthentication().getAuthorities().toArray()[0];
+        Event event = eventRepository.findById(eventId).orElseThrow(
                 () -> new ResponseStatusException( HttpStatus.NOT_FOUND , eventId + " Does not Exist !!!"));
-        eventRepository.deleteById(eventId);
+        if(role.getAuthority().contains("admin") || authentication.getName().contains(event.getBookingEmail())){
+            eventRepository.deleteById(eventId);
+        }else{
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"Access denied");
+        }
     }
 
-    public EventEditDto updateEvent(EventEditDto updateEvent, Integer eventId) {
+    public ResponseEntity<?> updateEvent(EventEditDto updateEvent, Integer eventId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Role role = (Role) SecurityContextHolder.getContext().getAuthentication().getAuthorities().toArray()[0];
         Event newEvent = modelMapper.map(updateEvent, Event.class);
-        Event event = eventRepository.findById(eventId).map(o -> mapEvent(o, newEvent)).orElseGet(() -> {
-            newEvent.setId(eventId);
-            return newEvent;
-        });
-        eventRepository.saveAndFlush(event);
-        return modelMapper.map(event, EventEditDto.class);
+        Event event = eventRepository.findById(eventId).orElseThrow(
+                () -> new ResponseStatusException( HttpStatus.NOT_FOUND , eventId + " Does not Exist !!!"));
+        if(role.getAuthority().contains("admin") || authentication.getName().contains(event.getBookingEmail())){
+            Event editEvent = eventRepository.findById(eventId).map(o -> mapEvent(o, newEvent)).orElseGet(() -> {
+                newEvent.setId(eventId);
+                return newEvent;
+            });
+            System.out.println(editEvent);
+            eventRepository.saveAndFlush(editEvent);
+            return ResponseEntity.ok(modelMapper.map(event, EventEditDto.class));
+        }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponse(HttpStatus.FORBIDDEN,"Access denied"));
     }
 
     private Event mapEvent(Event existingEvent, Event updateEvent) {
         if (updateEvent.getEventStartTime() != null) {
+            System.out.println(updateEvent.getEventStartTime());
             existingEvent.setEventStartTime(updateEvent.getEventStartTime());
         }
         if (updateEvent.getEventNotes() != null) {
+            System.out.println(updateEvent.getEventNotes());
             existingEvent.setEventNotes(updateEvent.getEventNotes());
         }
         return existingEvent;
